Feature cart

Approved-by: Karol Zborowski <krlzborowski@gmail.com>
Approved-by: Dariusz Stopka <dariusz.marek.stopka@gmail.com>

+++ database/migrations/2020_01_08_160841_create_events_table.php
@@ -21,7 +21,7 @@ class CreateEventsTable extends Migration
             $table->double('latitude', 8, 2);
             $table->double('longtitude', 8, 2);
             $table->bigInteger("places");
-            $table->bigInteger("price");
+            $table->float("price");
             $table->boolean('isFree');
             $table->timestamps();
         });
